split PermutationDistanceTests
